started to write a Test6 for Multiple Image difference

+++ Test/TEST_frameAllignment.h
@@ -16,7 +16,7 @@ typedef struct
     long    TotalTimeSn;
     float   fps;
     float   homograpyFoundPercent;
-    float   whitePixelPerFramePixels;
+    double  whitePixelPerFramePixels;
     QString HomographyMethod;
     QString usedDetector;
     QString usedDescriptor;
